refactor: refactor step 1
